feat: Preview panel seperation (#1021)
