+++ .gitignore
@@ -32,3 +32,6 @@ testsuite/.venv/
 
 # macOS-specific files
 .DS_Store
+
+# Agent configs
+CLAUDE.md
