+++ src/internal/handle_panel_navigation.go
@@ -50,23 +50,23 @@ func (m *model) createNewFilePanel(location string) error {
 		searchBar:        common.GenerateSearchBar(),
 	})
 
-	if m.fileModel.filePreview.open {
+	if m.fileModel.filePreview.IsOpen() {
 		// File preview panel width same as file panel
 		if common.Config.FilePreviewWidth == 0 {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth -
-				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1)
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth -
+				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1))
 		} else {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth)
 		}
 	}
 
 	m.fileModel.filePanels[m.filePanelFocusIndex].focusType = noneFocus
 	m.fileModel.filePanels[m.filePanelFocusIndex+1].focusType = returnFocusType(m.focusPanel)
-	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width -
+	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth() -
 		(4 + (len(m.fileModel.filePanels)-1)*2)) / len(m.fileModel.filePanels)
 	m.filePanelFocusIndex++
 
-	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width) / 20
+	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth()) / 20
 
 	for i := range m.fileModel.filePanels {
 		m.fileModel.filePanels[i].searchBar.Width = m.fileModel.width - 4
@@ -83,13 +83,13 @@ func (m *model) closeFilePanel() {
 	m.fileModel.filePanels = append(m.fileModel.filePanels[:m.filePanelFocusIndex],
 		m.fileModel.filePanels[m.filePanelFocusIndex+1:]...)
 
-	if m.fileModel.filePreview.open {
+	if m.fileModel.filePreview.IsOpen() {
 		// File preview panel width same as file panel
 		if common.Config.FilePreviewWidth == 0 {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth -
-				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1)
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth -
+				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1))
 		} else {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth)
 		}
 	}
 
@@ -97,11 +97,11 @@ func (m *model) closeFilePanel() {
 		m.filePanelFocusIndex--
 	}
 
-	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width -
+	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth() -
 		(4 + (len(m.fileModel.filePanels)-1)*2)) / len(m.fileModel.filePanels)
 	m.fileModel.filePanels[m.filePanelFocusIndex].focusType = returnFocusType(m.focusPanel)
 
-	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width) / 20
+	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth()) / 20
 
 	for i := range m.fileModel.filePanels {
 		m.fileModel.filePanels[i].searchBar.Width = m.fileModel.width - 4
@@ -109,22 +109,23 @@ func (m *model) closeFilePanel() {
 }
 
 func (m *model) toggleFilePreviewPanel() {
-	m.fileModel.filePreview.open = !m.fileModel.filePreview.open
-	m.fileModel.filePreview.width = 0
-	if m.fileModel.filePreview.open {
+	m.fileModel.filePreview.ToggleOpen()
+	m.fileModel.filePreview.SetWidth(0)
+	m.fileModel.filePreview.SetHeight(m.mainPanelHeight + 2)
+	if m.fileModel.filePreview.IsOpen() {
 		// File preview panel width same as file panel
 		if common.Config.FilePreviewWidth == 0 {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth -
-				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1)
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth -
+				(4 + (len(m.fileModel.filePanels))*2)) / (len(m.fileModel.filePanels) + 1))
 		} else {
-			m.fileModel.filePreview.width = (m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth
+			m.fileModel.filePreview.SetWidth((m.fullWidth - common.Config.SidebarWidth) / common.Config.FilePreviewWidth)
 		}
 	}
 
-	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width -
+	m.fileModel.width = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth() -
 		(4 + (len(m.fileModel.filePanels)-1)*2)) / len(m.fileModel.filePanels)
 
-	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.width) / 20
+	m.fileModel.maxFilePanel = (m.fullWidth - common.Config.SidebarWidth - m.fileModel.filePreview.GetWidth()) / 20
 
 	for i := range m.fileModel.filePanels {
 		m.fileModel.filePanels[i].searchBar.Width = m.fileModel.width - 4
