+++ src/internal/model_render.go
@@ -1,20 +1,11 @@
 package internal
 
 import (
-	"context"
-	"errors"
 	"fmt"
-	"image"
-	"io/fs"
 	"log/slog"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"slices"
-	"sort"
 	"strconv"
-	"strings"
-	"time"
 
 	"github.com/yorukot/superfile/src/internal/ui"
 	"github.com/yorukot/superfile/src/internal/ui/rendering"
@@ -22,10 +13,7 @@ import (
 	"github.com/yorukot/superfile/src/internal/common"
 	"github.com/yorukot/superfile/src/internal/utils"
 
-	"github.com/alecthomas/chroma/v2/lexers"
 	"github.com/charmbracelet/lipgloss"
-
-	"github.com/yorukot/ansichroma"
 
 	"github.com/yorukot/superfile/src/config/icon"
 )
@@ -57,7 +45,7 @@ func (m *model) filePanelRender() string {
 			m.fileModel.filePanels,
 		) != 0 &&
 			i == len(m.fileModel.filePanels)-1 {
-			if m.fileModel.filePreview.open {
+			if m.fileModel.filePreview.IsOpen() {
 				filePanelWidth = m.fileModel.width
 			} else {
 				filePanelWidth = (m.fileModel.width + (m.fullWidth-common.Config.SidebarWidth-
@@ -446,232 +434,4 @@ func (m *model) sortOptionsRender() string {
 
 func (m *model) filePreviewPanelRender() string {
 	return m.fileModel.filePreview.GetContent()
-}
-
-// Helper function to handle file info errors
-func (m *model) renderFileInfoError(r *rendering.Renderer, _ lipgloss.Style, err error) string {
-	slog.Error("Error get file info", "error", err)
-	return r.Render()
-}
-
-// Helper function to handle unsupported formats
-func (m *model) renderUnsupportedFormat(box lipgloss.Style) string {
-	return box.Render(common.FilePreviewUnsupportedFormatText)
-}
-
-// Helper function to handle unsupported mode
-func (m *model) renderUnsupportedFileMode(r *rendering.Renderer) string {
-	r.AddLines(common.FilePreviewUnsupportedFileMode)
-	return r.Render()
-}
-
-// Helper function to handle directory preview
-func (m *model) renderDirectoryPreview(r *rendering.Renderer, itemPath string, previewHeight int) string {
-	files, err := os.ReadDir(itemPath)
-	if err != nil {
-		slog.Error("Error render directory preview", "error", err)
-		r.AddLines(common.FilePreviewDirectoryUnreadableText)
-		return r.Render()
-	}
-
-	if len(files) == 0 {
-		r.AddLines(common.FilePreviewEmptyText)
-		return r.Render()
-	}
-
-	sort.Slice(files, func(i, j int) bool {
-		if files[i].IsDir() && !files[j].IsDir() {
-			return true
-		}
-		if !files[i].IsDir() && files[j].IsDir() {
-			return false
-		}
-		return files[i].Name() < files[j].Name()
-	})
-
-	for i := 0; i < previewHeight && i < len(files); i++ {
-		file := files[i]
-		style := common.GetElementIcon(file.Name(), file.IsDir(), common.Config.Nerdfont)
-		res := lipgloss.NewStyle().Foreground(lipgloss.Color(style.Color)).Background(common.FilePanelBGColor).
-			Render(style.Icon+" ") + common.FilePanelStyle.Render(file.Name())
-		r.AddLines(res)
-	}
-	return r.Render()
-}
-
-// Helper function to handle image preview
-func (m *model) renderImagePreview(box lipgloss.Style, itemPath string, previewWidth,
-	previewHeight int, sideAreaWidth int) string {
-	if !m.fileModel.filePreview.open {
-		return box.Render("\n --- Preview panel is closed ---")
-	}
-
-	if !common.Config.ShowImagePreview {
-		return box.Render("\n --- Image preview is disabled ---")
-	}
-
-	// Use the new auto-detection function to choose the best renderer
-	imageRender, err := m.imagePreviewer.ImagePreview(itemPath, previewWidth, previewHeight,
-		common.Theme.FilePanelBG, sideAreaWidth)
-	if errors.Is(err, image.ErrFormat) {
-		return box.Render("\n --- " + icon.Error + " Unsupported image formats ---")
-	}
-
-	if err != nil {
-		slog.Error("Error convert image to ansi", "error", err)
-		return box.Render("\n --- " + icon.Error + " Error convert image to ansi ---")
-	}
-
-	// Check if this looks like Kitty protocol output (starts with escape sequences)
-	// For Kitty protocol, avoid using lipgloss alignment to prevent layout drift
-	if strings.HasPrefix(imageRender, "\x1b_G") {
-		rendered := common.FilePreviewBox(previewHeight, previewWidth).Render(imageRender)
-		return rendered
-	}
-
-	// For ANSI output, we can safely use vertical alignment
-	return box.AlignVertical(lipgloss.Center).AlignHorizontal(lipgloss.Center).Render(imageRender)
-}
-
-// Helper function to handle text file preview
-func (m *model) renderTextPreview(r *rendering.Renderer, box lipgloss.Style, itemPath string,
-	previewWidth, previewHeight int) string {
-	format := lexers.Match(filepath.Base(itemPath))
-	if format == nil {
-		isText, err := common.IsTextFile(itemPath)
-		if err != nil {
-			slog.Error("Error while checking text file", "error", err)
-			return box.Render(common.FilePreviewError)
-		} else if !isText {
-			return box.Render(common.FilePreviewUnsupportedFormatText)
-		}
-	}
-
-	fileContent, err := utils.ReadFileContent(itemPath, previewWidth, previewHeight)
-	if err != nil {
-		slog.Error("Error open file", "error", err)
-		return box.Render(common.FilePreviewError)
-	}
-
-	if fileContent == "" {
-		return box.Render(common.FilePreviewEmptyText)
-	}
-
-	if format != nil {
-		background := ""
-		if !common.Config.TransparentBackground {
-			background = common.Theme.FilePanelBG
-		}
-		if common.Config.CodePreviewer == "bat" {
-			if batCmd == "" {
-				return box.Render("\n --- " + icon.Error +
-					" 'bat' is not installed or not found. ---\n --- Cannot render file preview. ---")
-			}
-			fileContent, err = getBatSyntaxHighlightedContent(itemPath, previewHeight, background)
-		} else {
-			fileContent, err = ansichroma.HightlightString(fileContent, format.Config().Name,
-				common.Theme.CodeSyntaxHighlightTheme, background)
-		}
-		if err != nil {
-			slog.Error("Error render code highlight", "error", err)
-			return box.Render("\n" + common.FilePreviewError)
-		}
-	}
-
-	r.AddLines(fileContent)
-	return r.Render()
-}
-func (m *model) filePreviewPanelRenderWithDimensions() string {
-	panel := m.getFocusedFilePanel()
-	if len(panel.element) == 0 {
-		return m.fileModel.filePreview.RenderText("")
-	}
-	itemPath := panel.getSelectedItem().location
-	return m.filePreviewPanelRenderWithPath(itemPath)
-}
-
-// TODO: Avoid accessing the model here.
-// There is used in tea.Cmd for async rendering. Avoid access inside async
-// command closures to reduce data-race risk. Require immutable values as parameter
-// (e.g., previewW, previewH, sideAreaWidth, imagePreviewer, theme colors) and pass
-// them to a pure renderer that doesn’t touch m.
-func (m *model) filePreviewPanelRenderWithPath(itemPath string) string {
-	previewHeight := m.fileModel.filePreview.height
-	previewWidth := m.fileModel.filePreview.width
-
-	box := common.FilePreviewBox(previewHeight, previewWidth)
-	r := ui.FilePreviewPanelRenderer(previewHeight, previewWidth)
-	clearCmd := m.imagePreviewer.ClearKittyImages()
-
-	// This could create errors if panel.cursor ever becomes negative, or goes out of bounds
-	// We should have a panel validation function in our View() function
-	// Panel is a full fledged object with own state, its accessed and modified so many times.
-	// Ideally we dont should never access data from it via directly accessing its variables
-	// TODO : Instead we should have helper functions for panel object and access data that way
-	// like panel.GetCurrentSelectedElem() . This abstration of implemetation of panel is needed.
-	// Now this lack of abstraction has caused issues ( See PR#730 ) . And now
-	// someone needs to scan through the entire codebase to figure out which access of panel
-	// data is causing crash.
-
-	fileInfo, infoErr := os.Stat(itemPath)
-	if infoErr != nil {
-		return m.renderFileInfoError(r, box, infoErr) + clearCmd
-	}
-	slog.Debug("Attempting to render preview", "itemPath", itemPath,
-		"mode", fileInfo.Mode().String(), "isRegular", fileInfo.Mode().IsRegular())
-
-	// For non regular files which are not directories Dont try to read them
-	// See Issue
-	if !fileInfo.Mode().IsRegular() && (fileInfo.Mode()&fs.ModeDir) == 0 {
-		return m.renderUnsupportedFileMode(r) + clearCmd
-	}
-
-	ext := filepath.Ext(itemPath)
-	if slices.Contains(common.UnsupportedPreviewFormats, ext) {
-		return m.renderUnsupportedFormat(box) + clearCmd
-	}
-
-	if fileInfo.IsDir() {
-		return m.renderDirectoryPreview(r, itemPath, previewHeight) + clearCmd
-	}
-
-	if isImageFile(itemPath) {
-		return m.renderImagePreview(box, itemPath, previewWidth, previewHeight, m.fullWidth-previewWidth+1)
-	}
-
-	return m.renderTextPreview(r, box, itemPath, previewWidth, previewHeight) + clearCmd
-}
-
-func getBatSyntaxHighlightedContent(itemPath string, previewLine int, background string) (string, error) {
-	// --plain: use the plain style without line numbers and decorations
-	// --force-colorization: force colorization for non-interactive shell output
-	// --line-range <:m>: only read from line 1 to line "m"
-	batArgs := []string{itemPath, "--plain", "--force-colorization", "--line-range", fmt.Sprintf(":%d", previewLine-1)}
-
-	// set timeout for the external command execution to 500ms max
-	ctx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
-	defer cancel()
-
-	cmd := exec.CommandContext(ctx, batCmd, batArgs...)
-
-	fileContentBytes, err := cmd.Output()
-	if err != nil {
-		slog.Error("Error render code highlight", "error", err)
-		return "", err
-	}
-
-	fileContent := string(fileContentBytes)
-	if !common.Config.TransparentBackground {
-		fileContent = setBatBackground(fileContent, background)
-	}
-	return fileContent, nil
-}
-
-func setBatBackground(input string, background string) string {
-	tokens := strings.Split(input, "\x1b[0m")
-	backgroundStyle := lipgloss.NewStyle().Background(lipgloss.Color(background))
-	for idx, token := range tokens {
-		tokens[idx] = backgroundStyle.Render(token)
-	}
-	return strings.Join(tokens, "\x1b[0m")
 }
